CA-705 Ensure some delay even when running in fast mode
When running in fast mode the run at was set to the current time. This
meant that the worker running the job picked up the new job immediatly.
If you have a job that runs quickly this can result in log and job spam.
In order to prevent this and to better honour the back off time in the
delayed job polling time by adding a second to the run at ensures the
job wont get picked up immediatly.