Ensure we can schedule substrings of jobs
When checking to see if a job exists we should allow a job to be
scheduled that is a substring of another job. Previously if we tried to
scheduled a job of class Test after scheduling a job of class UnderTest
the job would not be scheduled. This commit fixes this case.

Note as we have no backend the spec I wrote for this cannot run. At some
stage we could look to adding an in memory db as a dev dependency but
this fix is needed right now.